fix(file-provider): Logging now adopts category string in unified logging output, too
